Several fixes in one commit (sorry):
Debugger now refers to the stack pointer register as SP rather than the old S, and we added the word S, an alias to STEP.

The assembler now allow referring to the stack pointer as either S or SP.

These two were to make the debugger consistent with the 8086 debugger name for the stack pointer.

Numerous bug fixes in the screens vocabulary.

In the editor, [re]search now skips over blank or binary screens.

	modified:   atari/FF.scn