Smoother overlay loading on screen 3.
Numerous substitutions of 0 # <reg> movq, for <reg> clr, Apparently, this is faster and takes the same space (2 bytes). See screen 32.

Some assembler words still need trailing ;c to stop assembly.

A tool to debug file directory entry names, LH, by hex dumping the entry for the file name. Useful for detecting embedded control characters, etc.

The line editor now moves to the next space in the current line with ctl left and right arrow. N.B. Known to work on Hatari, not necessarily on other platforms.

Updated the line editor help, and added the PC keys for Hatari to it.

Added a small coding challenge solution: check that a character is a valid digit in the current base.

Some target compiler improvements.

	modified:   atari/FF.scn
	modified:   atari/FFX.scn